Switch HTTP/HTTPS proxy from connect to using AIM-HTTP gate
git-svn-id: file:///.data/svn_convert/svn@477 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ trunk/sim/libicq/service.cpp
@@ -309,7 +309,7 @@ void ICQClient::setShare(bool bState)
     b.pack((char*)(ICQClientPrivate::SHARED_FILES_SIGN), 16);
     b << (char)4 << (unsigned short)1;
     b.pack((unsigned long)(bState ? 1 : 0));
-    b.pack(now);
+    b.pack((unsigned long)now);
     b.pack((unsigned long)0);
     b.pack((unsigned short)1);
     p->sendUpdate(b);
